# api_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blocknet Daemon API",
    "description": "Authenticated JSON API for the Blocknet privacy blockchain daemon. Used by GUI wallets (Tauri) and tooling. All private endpoints require Bearer token authentication via a cookie file (`<dataDir>/api.cookie`) written on daemon startup. Public routes (GET-only blockchain queries) are also served unauthenticated when using `--explorer` mode.",
    "version": "0.1.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8332",
      "description": "Local daemon (default --api binding)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/block/{id}": {
      "get": {
        "operationId": "getBlock",
        "tags": ["Blockchain"],
        "summary": "Get block by height or hash",
        "description": "Returns a block by numeric height or 64-character hex hash. Includes transaction summaries and confirmation count.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Block height (integer) or block hash (64 hex characters)",
            "schema": {
              "type": "string"
            },
            "examples": {
              "height": {
                "summary": "By height",
                "value": "14207"
              },
              "hash": {
                "summary": "By hash",
                "value": "0000c3a5b7e2d1f489a1bc37e5d204f8c612aa9b33e7f04d12b8a9e6c7f50321"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Block data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                },
                "example": {
                  "height": 14207,
                  "hash": "0000c3a5b7e2d1f489a1bc37e5d204f8c612aa9b33e7f04d12b8a9e6c7f50321",
                  "prev_hash": "0000a1f38e2bc9d7a436ef21c87b3d4e59a0f1b267e3c9d58a2b41f6e7083c19",
                  "merkle_root": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
                  "timestamp": 1738800120,
                  "difficulty": 100000,
                  "nonce": 847291,
                  "tx_count": 2,
                  "transactions": [
                    {
                      "hash": "c7f2e1d3a4b596870123456789abcdef0123456789abcdef0123456789abcdef",
                      "is_coinbase": true,
                      "inputs": 0,
                      "outputs": 1,
                      "fee": 0
                    },
                    {
                      "hash": "a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2",
                      "is_coinbase": false,
                      "inputs": 1,
                      "outputs": 2,
                      "fee": 15000
                    }
                  ],
                  "confirmations": 1,
                  "reward": 72325093035
                }
              }
            }
          },
          "400": {
            "description": "Invalid block ID format",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "id must be a height or 64-char hex hash" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "Block not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "block not found" }
              }
            }
          }
        }
      }
    },
...
    "/api/peers/banned": {
      "get": {
        "operationId": "getBannedPeers",
        "tags": ["Network"],
        "summary": "Banned peers",
        "description": "Returns the list of banned peers with ban reason, count, permanence, and expiry time.",
        "responses": {
          "200": {
            "description": "Banned peer list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "banned": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/BanEntry" }
                    }
                  }
                },
                "example": {
                  "count": 2,
                  "banned": [
                    {
                      "peer_id": "12D3KooWMalicious111111111111111111111111111111111111",
                      "reason": "sent invalid block",
                      "ban_count": 3,
                      "permanent": false,
                      "expires_at": "2026-02-06T19:30:00Z"
                    },
                    {
                      "peer_id": "12D3KooWMalicious222222222222222222222222222222222222",
                      "reason": "excessive invalid transactions",
                      "ban_count": 5,
                      "permanent": true
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/mining/start": {
      "post": {
        "operationId": "startMining",
        "tags": ["Mining"],
        "summary": "Start mining",
        "description": "Starts the Argon2id proof-of-work miner. Each thread allocates ~2GB RAM. No request body required.",
        "responses": {
          "200": {
            "description": "Mining started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "running": { "type": "boolean" }
                  }
                },
                "example": { "running": true }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": {
            "description": "Mining already running",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "mining already running" }
              }
            }
          }
        }
      }
    },
    "/api/mining/stop": {
      "post": {
        "operationId": "stopMining",
        "tags": ["Mining"],
        "summary": "Stop mining",
        "description": "Stops the miner. No request body required.",
        "responses": {
          "200": {
            "description": "Mining stopped",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "running": { "type": "boolean" }
                  }
                },
                "example": { "running": false }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": {
            "description": "Mining not running",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "mining not running" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token read from `<dataDir>/api.cookie`. Generated fresh on every daemon startup using `crypto/rand` (32 bytes, hex-encoded = 64 chars). File has 0600 permissions. Deleted on graceful shutdown."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        },
        "example": { "error": "block not found" }
      },
...
      "Block": {
        "type": "object",
        "properties": {
          "height": { "type": "integer" },
          "hash": { "type": "string", "description": "Full block hash (64 hex chars)" },
          "prev_hash": { "type": "string", "description": "Previous block hash (64 hex chars)" },
          "merkle_root": { "type": "string", "description": "Merkle root of transactions (64 hex chars)" },
          "timestamp": { "type": "integer", "description": "Block timestamp (Unix seconds)" },
          "difficulty": { "type": "integer", "description": "Block difficulty target" },
          "nonce": { "type": "integer", "description": "PoW nonce" },
          "tx_count": { "type": "integer", "description": "Number of transactions in the block" },
          "transactions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TxSummary" },
            "description": "Transaction summaries (first tx is coinbase for height > 0)"
          },
          "confirmations": { "type": "integer", "description": "Confirmations since this block (chain_height - block_height + 1)" },
          "reward": { "type": "integer", "description": "Block reward in atomic units (smooth exponential decay to 0.2 BNT tail)" }
        },
        "example": {
          "height": 14207,
          "hash": "0000c3a5b7e2d1f489a1bc37e5d204f8c612aa9b33e7f04d12b8a9e6c7f50321",
          "prev_hash": "0000a1f38e2bc9d7a436ef21c87b3d4e59a0f1b267e3c9d58a2b41f6e7083c19",
          "merkle_root": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
          "timestamp": 1738800120,
          "difficulty": 100000,
          "nonce": 847291,
          "tx_count": 2,
          "transactions": [
            {
              "hash": "c7f2e1d3a4b596870123456789abcdef0123456789abcdef0123456789abcdef",
              "is_coinbase": true,
              "inputs": 0,
              "outputs": 1,
              "fee": 0
            },
            {
              "hash": "a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2",
              "is_coinbase": false,
              "inputs": 1,
              "outputs": 2,
              "fee": 15000
            }
          ],
          "confirmations": 1,
          "reward": 72325093035
        }
      },
      "TxSummary": {
        "type": "object",
        "properties": {
          "hash": { "type": "string", "description": "Transaction hash (64 hex chars)" },
          "is_coinbase": { "type": "boolean", "description": "True for block reward transaction" },
          "inputs": { "type": "integer", "description": "Number of inputs (0 for coinbase)" },
          "outputs": { "type": "integer", "description": "Number of outputs" },
          "fee": { "type": "integer", "description": "Fee in atomic units (0 for coinbase)" }
        },
        "example": {
          "hash": "a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2",
          "is_coinbase": false,
          "inputs": 1,
          "outputs": 2,
          "fee": 15000
        }
      },
...
      "BanEntry": {
        "type": "object",
        "properties": {
          "peer_id": { "type": "string", "description": "libp2p peer ID" },
          "reason": { "type": "string", "description": "Human-readable ban reason" },
          "ban_count": { "type": "integer", "description": "Number of times this peer has been banned" },
          "permanent": { "type": "boolean", "description": "True if ban does not expire" },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 expiry timestamp (omitted when permanent)"
          }
        },
        "example": {
          "peer_id": "12D3KooWN7z3vEqr4C8pMk2axGfRbLVJzqPEasiY4L9DMqFjVnGq",
          "reason": "sent invalid block",
          "ban_count": 3,
          "permanent": false,
          "expires_at": "2026-02-06T19:30:00Z"
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid `Authorization: Bearer <token>` header",
        "content": {
          "text/plain": {
            "schema": { "type": "string" },
            "example": "unauthorized"
          }
        }
      },
...
    }
  },
...
}